CB-12895 Move shared connectors to common module
