+++ cloud-aws-cloudformation/build.gradle
@@ -30,8 +30,6 @@ dependencies {
   compile group: 'org.freemarker', name: 'freemarker', version: freemarkerVersion
   compile group: 'commons-net', name: 'commons-net', version: '3.6'
 
-  implementation group: 'com.dyngr', name: 'polling', version: '1.1.3'
-
   testCompile group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: springBootVersion
   testCompile group: 'org.springframework.boot', name: 'spring-boot-starter-freemarker', version: springBootVersion
   testCompile group: 'org.powermock', name: 'powermock-module-junit4', version: powermockVersion
