+++ cloud-aws-common/build.gradle
@@ -17,6 +17,9 @@ dependencies {
   compile project(':cloud-common')
   compile project(':cloud-reactor-api')
   compile project(':template-manager-tag')
+  compile project(':auth-connector')
+
+  implementation group: 'com.dyngr', name: 'polling', version: '1.1.3'
 
   compile(group: 'com.amazonaws', name: 'aws-java-sdk-core', version: awsSdkVersion) {
     exclude group: 'commons-logging'
