+++ cloud-aws-common/src/test/java/com/sequenceiq/cloudbreak/cloud/aws/common/validator/AwsRangerAuditRolePermissionValidatorTest.java
@@ -1,4 +1,4 @@
-package com.sequenceiq.cloudbreak.cloud.aws.validator;
+package com.sequenceiq.cloudbreak.cloud.aws.common.validator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -10,39 +10,40 @@ import com.sequenceiq.cloudbreak.cloud.model.filesystem.CloudS3View;
 import com.sequenceiq.cloudbreak.service.identitymapping.AccountMappingSubject;
 import com.sequenceiq.common.api.cloudstorage.StorageLocationBase;
 import com.sequenceiq.common.model.CloudIdentityType;
+import com.sequenceiq.common.model.CloudStorageCdpService;
 import com.sequenceiq.common.model.FileSystemType;
 
-public class AwsDataAccessRolePermissionValidatorTest extends AwsIDBrokerMappedRolePermissionValidatorTest {
-    private final AwsDataAccessRolePermissionValidator awsDataAccessRolePermissionValidator =
-            new AwsDataAccessRolePermissionValidator();
+public class AwsRangerAuditRolePermissionValidatorTest extends AwsIDBrokerMappedRolePermissionValidatorTest {
+    private final AwsRangerAuditRolePermissionValidator awsRangerAuditRolePermissionValidator =
+            new AwsRangerAuditRolePermissionValidator();
 
     @Override
     public AwsIDBrokerMappedRolePermissionValidator getValidator() {
-        return awsDataAccessRolePermissionValidator;
+        return awsRangerAuditRolePermissionValidator;
     }
 
     @Override
     public void testGetUsers() {
-        assertThat(awsDataAccessRolePermissionValidator.getUsers())
-                .isEqualTo(AccountMappingSubject.DATA_ACCESS_USERS);
+        assertThat(awsRangerAuditRolePermissionValidator.getUsers())
+                .isEqualTo(AccountMappingSubject.RANGER_AUDIT_USERS);
     }
 
     @Override
     public void testGetPolicyFileNames() {
         List<String> expectedPolicyFileNamesNoS3Guard = Arrays.asList(
                 "aws-cdp-bucket-access-policy.json",
-                "aws-cdp-datalake-admin-s3-policy.json"
+                "aws-cdp-ranger-audit-s3-policy.json"
         );
-        List<String> policyFileNamesNoS3Guard = awsDataAccessRolePermissionValidator
+        List<String> policyFileNamesNoS3Guard = awsRangerAuditRolePermissionValidator
                 .getPolicyFileNames(false);
         assertThat(policyFileNamesNoS3Guard).isEqualTo(expectedPolicyFileNamesNoS3Guard);
 
         List<String> expectedPolicyFileNamesS3Guard = Arrays.asList(
                 "aws-cdp-bucket-access-policy.json",
-                "aws-cdp-datalake-admin-s3-policy.json",
+                "aws-cdp-ranger-audit-s3-policy.json",
                 "aws-cdp-dynamodb-policy.json"
         );
-        List<String> policyFileNamesS3Guard = awsDataAccessRolePermissionValidator
+        List<String> policyFileNamesS3Guard = awsRangerAuditRolePermissionValidator
                 .getPolicyFileNames(true);
         assertThat(policyFileNamesS3Guard).isEqualTo(expectedPolicyFileNamesS3Guard);
     }
@@ -50,16 +51,24 @@ public class AwsDataAccessRolePermissionValidatorTest extends AwsIDBrokerMappedR
     @Override
     public void testGetStorageLocationBase() {
         String path = "testBucket/ranger/audit";
-        String expectedStorageLocationBase = "testBucket/ranger/audit";
+        String expectedStorageLocationBase = "testBucket";
         StorageLocationBase location = new StorageLocationBase();
         location.setValue(String.format("%s://%s", FileSystemType.S3.getProtocol(), path));
-        String storageLocationBase = awsDataAccessRolePermissionValidator.getStorageLocationBase(location);
+        String storageLocationBase = awsRangerAuditRolePermissionValidator.getStorageLocationBase(location);
         assertThat(storageLocationBase).isEqualTo(expectedStorageLocationBase);
     }
 
     @Override
     public void testCheckLocation() {
-        assertThat(awsDataAccessRolePermissionValidator.checkLocation(new StorageLocationBase())).isTrue();
+        assertThat(awsRangerAuditRolePermissionValidator.checkLocation(new StorageLocationBase())).isFalse();
+
+        StorageLocationBase nonRangerAuditLocation = new StorageLocationBase();
+        nonRangerAuditLocation.setType(CloudStorageCdpService.HBASE_ROOT);
+        assertThat(awsRangerAuditRolePermissionValidator.checkLocation(nonRangerAuditLocation)).isFalse();
+
+        StorageLocationBase rangerAuditLocation = new StorageLocationBase();
+        rangerAuditLocation.setType(CloudStorageCdpService.RANGER_AUDIT);
+        assertThat(awsRangerAuditRolePermissionValidator.checkLocation(rangerAuditLocation)).isTrue();
     }
 
     @Override
@@ -78,7 +87,7 @@ public class AwsDataAccessRolePermissionValidatorTest extends AwsIDBrokerMappedR
         storageLocationBase.setValue(storageLocationBaseStr);
         CloudS3View cloudFileSystem = new CloudS3View(CloudIdentityType.ID_BROKER);
         cloudFileSystem.setS3GuardDynamoTableName(dynamodbTableName);
-        Map<String, String> policyJsonReplacements = awsDataAccessRolePermissionValidator
+        Map<String, String> policyJsonReplacements = awsRangerAuditRolePermissionValidator
                 .getPolicyJsonReplacements(storageLocationBase,
                         cloudFileSystem);
 
@@ -99,7 +108,7 @@ public class AwsDataAccessRolePermissionValidatorTest extends AwsIDBrokerMappedR
         StorageLocationBase storageLocationBase = new StorageLocationBase();
         storageLocationBase.setValue(storageLocationBaseStr);
         CloudS3View cloudFileSystem = new CloudS3View(CloudIdentityType.ID_BROKER);
-        Map<String, String> policyJsonReplacements = awsDataAccessRolePermissionValidator
+        Map<String, String> policyJsonReplacements = awsRangerAuditRolePermissionValidator
                                                         .getPolicyJsonReplacements(storageLocationBase,
                                                             cloudFileSystem);
